reentrant deserialization

+++ src/editor/editor_server.cpp
@@ -168,7 +168,7 @@ struct EditorServerImpl
 		Entity& getSelectedEntity() { return m_selected_entity; }
 		bool isGameMode() const { return m_is_game_mode; }
 		void save(FS::IFile& file, const char* path);
-		void load(FS::IFile& file, const char* path);
+		void load(FS::IFile& file, const char* path, bool reset_universe);
 		void onMessage(void* msgptr, int size);
 		EditorIconHit raycastEditorIcons(const Vec3& origin, const Vec3& dir);
 
@@ -491,7 +491,7 @@ void EditorServerImpl::stopGameMode()
 	m_is_game_mode = false;
 	m_engine.getScriptSystem().stop();
 	m_game_mode_file->seek(FS::SeekMode::BEGIN, 0);
-	load(*m_game_mode_file, "GameMode");
+	load(*m_game_mode_file, "GameMode", false);
 	m_engine.getFileSystem().close(m_game_mode_file);
 	m_game_mode_file = NULL;
 }
@@ -652,7 +652,7 @@ void EditorServerImpl::loadMap(FS::IFile* file, bool success, FS::FileSystem& fs
 	ASSERT(success);
 	if(success)
 	{
-		load(*file, "unknown map"); /// TODO file path
+		load(*file, "unknown map", true); /// TODO file path
 	}
 
 	fs.close(file);
@@ -666,11 +666,14 @@ void EditorServerImpl::newUniverse()
 }
 
 
-void EditorServerImpl::load(FS::IFile& file, const char* path)
+void EditorServerImpl::load(FS::IFile& file, const char* path, bool reset_universe)
 {
 	g_log_info.log("editor server", "parsing universe...");
-	destroyUniverse();
-	createUniverse(false);
+	if (reset_universe)
+	{
+		destroyUniverse();
+		createUniverse(false);
+	}
 	JsonSerializer serializer(file, JsonSerializer::READ, path);
 	m_engine.deserialize(serializer);
 	m_camera = m_engine.getRenderScene()->getCameraInSlot("editor").entity;
